aio/src/app/shared/theme-picker/theme-storage/theme-storage.service.ts
import {Injectable} from '@angular/core';

export interface DocsSiteTheme {
  name: string;
  displayName?: string;
  accent: string;
  primary: string;
  isDark?: boolean;
  icon: string;
  isDefault?: boolean;
}


@Injectable()
export class ThemeStorageService {
  static storageKey = 'docs-theme-storage-current-name';

  storeTheme(theme: DocsSiteTheme) {
    try {
      window.localStorage[ThemeStorageService.storageKey] = theme.name;
    } catch { }
  }

  getStoredThemeName(): string | null {
    try {
      return window.localStorage[ThemeStorageService.storageKey] || null;
    } catch {
      return null;
    }
  }

  clearStorage() {
    try {
      window.localStorage.removeItem(ThemeStorageService.storageKey);
    } catch { }
  }
}
